turb_pred/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.3
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.5.2
joblib==1.4.2
psycopg2-binary==2.9.10
